// src/pages/auth/types.tsx
export const TeacherSpecification = {
    IELTS: 'IELTS',
    SPEAKING: 'SPEAKING',
    GRAMMAR: 'GRAMMAR',
    BUSINESS: 'BUSINESS',
    KIDS: 'KIDS',
} as const;

export type TeacherSpecification = typeof TeacherSpecification[keyof typeof TeacherSpecification];
// ...
export type SortField = 'fullName' | 'rating' | 'createdAt' | 'hourPrice';
export type SortOrder = 'ASC' | 'DESC';

export interface TeacherFilters {
    search?: string;
    specification?: TeacherSpecification | '';
    level?: string;
    minRating?: number;
    maxRating?: number;
    minPrice?: number;
    maxPrice?: number;
    isActive?: boolean;
    sortBy?: SortField;
    sortOrder?: SortOrder;
    page?: number;
    limit?: number;
}